+++ app/backend/src/couchers/servicers/reporting.py
@@ -3,29 +3,32 @@ import requests
 
 from couchers.config import config
 from couchers.db import session_scope
-from couchers.models import User
+from couchers.models import User, ContentReport
 from couchers.sql import couchers_select as select
 from proto import reporting_pb2_grpc
 from task import send_content_reporting_email
 
 
 class Reporting(reporting_pb2_grpc.ReportingServicer):
-    def _version(self):
-        return config["VERSION"]
+    def ContentReport(self, request, context):
+        with session_scope() as session:
+            user = session.execute(
+                select(User).where(User.id == context.user_id)
+            ).scalar_one()
+            content_report = ContentReport(
+                subject=request.subject,
+                content_ref=request.content_ref,
+                content_owner_user_id=request.content_owner_user_id,
+                description=request.description,
+                user_id=context.user_id,
+                user_agent=request.user_agent,
+                page=request.page
+            )
 
-    def Version(self, request, context):
-        return reporting_pb2_grpc.VersionInfo(version=self._version())
+            session.add(ContentReport)
 
-    def Report(self, request, context):
-        if context.user_id:
-            with session_scope() as session:
-                username = session.execute(select(User.username).where(User.id == context.user_id)).scalar_one()
-                user_details = f"{username} ({context.user_id})"
-        else:
-            user_details = "<not logged in>"
+            session.flush()
 
-        send_content_reporting_email()
+            send_content_reporting_email(content_report)
 
-        return reporting_pb2_grpc.ReportBugRes(
-            report_id=f"#{issue_number}", report_url=f"https://github.com/{repo}/issues/{issue_number}"
-        )
+            return empty_pb2.Empty()
